From Paul Martz, added support to textured light points into osgSim and OpenFlight loader, as part of the OpenFlight v16.0 spec.

+++ src/osgPlugins/OpenFlight/PaletteRecords.cpp
@@ -502,7 +502,8 @@ protected:
         appearance->fadeOutDuration = in.readFloat32();
         appearance->LODRangeRatio = in.readFloat32();
         appearance->LODScale = in.readFloat32();
-        appearance->texturePatternIndex = in.readInt32(-1);
+        appearance->texturePatternIndex = in.readInt16(-1);
+        // The final short is reserved; don't bother reading it.
 
         // Add to pool
         LightPointAppearancePool* lpaPool = document.getOrCreateLightPointAppearancePool();
